# docs/swagger.yaml
definitions:
...
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      kind:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /booking-assigned/transfer:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view booking
        transfer that assigned and whose are not accepted by the driver
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List Booking Transfer Assigned
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/accept/{id}:
    get:
      consumes:
      - application/json
      description: A POST Request for Driver to Accept specific booking transfer by
        it's ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking Transfer Assigned ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to accept booking transfer
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/accepted:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view booking
        transfer that assigned and whose are not accepted by the driver
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List Booking Transfer Assigned
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/cancel:
    get:
      consumes:
      - application/json
      description: A POST Request for Driver to Cancel specific booking transfer by
        it's ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking Transfer Assigned ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to cancel booking transfer request
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/cancelled:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view booking
        transfer that assigned and whose are not accepted by the driver
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List Booking Transfer Assigned
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/completed:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view booking
        transfer that assigned and whose are not accepted by the driver
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List Booking Transfer Assigned
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/ongoing:
    get:
      consumes:
      - application/json
      description: A POST Request for Driver to set Complete Pickup specific booking
        transfer by it's ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking Transfer Assigned ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set status to Complete PickUp and notify passenger
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/ongoing/{id}:
    get:
      consumes:
      - application/json
      description: A POST Request for Driver to set status to be "OnGoing" to specific
        booking transfer by it's ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking Transfer Assigned ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set status to ongoing and notify passenger that driver on
        it's way to pickup
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/switch/{id}/{driverId}/{plateNumber}:
    get:
      consumes:
      - application/json
      description: A POST Request for Driver to set Complete Pickup specific booking
        transfer by it's ID
      parameters:
      - description: BookingTransferAssigned ID
        in: path
        name: id
        required: true
        type: string
      - description: DriverID ID
        in: path
        name: driverId
        required: true
        type: string
      - description: Car Plate Number
        in: path
        name: plateNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set status to Complete PickUp and notify passenger
      tags:
      - Booking Assgined - Transfer
...
swagger: "2.0"